Spatial overlap

--

findOverlap(
  object,
  of_sample = "",
  input,
  method_gs = "zscore",
  method_ovl = "bayesian",
  pt_clrp = "milo",
  pt_clrsp = "inferno",
  pt_size = 2,
  pt_alpha = 0.9,
  verbose = TRUE
)

Arguments

object
A valid spata-object.

of_sample
The sample(s) of interest specified as a single character value or vector. If set to "" (the default) the first sample is chosen.

input
Character vector of length 2. Specifies the genes or gene-sets whose overlap you want to calculate.

method_gs
Character value. The method according to which gene sets will be handled specified as a character of length one. This can be either 'mean' or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().

method_ovl
Character value. One of 'classic', 'bayesian'. Decides according to which method the spatial overlap is calculated.

pt_clrp
The color panel to be used if the specified variable displayed by color is categorical/discrete. Run all_colorpanels() to see valid input.

pt_clrsp
The color spectrum to be used if the specified variable displayed by color is continuous. Run all_colorspectra() to see valid input..

pt_size
The size of the points specified as a single numeric value.

pt_alpha
The transparency of the points specified as single numeric value.

verbose
Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

Value

A named list containing the results (plots, estimated correlation value, data.frames).