Plot differentially expressed genes

Takes the results from your de-analysis and uses the expression information of your spata-object to plot a heatmap displaying the differentially expressed genes of every cluster.

plotDeHeatmap(
  object,
  of_sample = "",
  de_df,
  across,
  across_subset = NULL,
  n_barcode_spots = 100,
  verbose = TRUE,
  hm_colors = viridis::viridis(15),
  ...
)

Arguments

object
A valid spata-object.

of_sample
The sample(s) of interest specified as a single character value or vector. If set to "" (the default) the first sample is chosen.

de_df
A data.frame containing information about differentially expressed genes. Must contain the variables:

gene

Character. The differentially expressed genes.

cluster

Character. The clusters (or experimental groups) across which the analysis was performed.

avg_logFC

Numeric. The average log-fold change to which the belonging gene was differentially expressed..

p_val

Numeric. The p-values.

p_val_adj

Numeric. The adjusted p-values.

Hint: Use the resulting data.frame of SPATA::findDE() or it's descendants as input.

across
Character value. The name of the discrete feature-variable of interest.

Hint: Character- or factor- variables are discrete variables. These functionally assign the barcode spots to distinct groups or clusters (e.g. segment or seurat_clusters) whoose properties you might want to compare against each other. Use getFeatureNames() to get an overview of the features variables your spata-object contains.

across_subset
Character vector or NULL. Specify the particular groups or clusters of interest the feature-variable specified in argument across contains. If set to NULL all of them are chosen.

Hint: Use getFeatureValues() to obtain all available groups of a certain feature-variable.

n_barcode_spots
The number of barcode-spots belonging to each cluster you want to include in the matrix. Should be lower than the total number of barcode-spots of every cluster and can be deployed in order to keep the heatmap clear and aesthetically pleasing.

verbose
Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

hm_colors
A vector of colors to be used.

...
Additional parameters given to pheatmap::pheatmap().

Value

A heatmap of class 'pheatmap'.