Plot several surface plots at the same time

Plots a surface plot for every valid element in argument variables and colors it according to each element.

plotSurfaceComparison() Takes the spata-object as the starting point and creates the necessary data.frame from scratch according to additional parameters.

plotSurfaceComparison2() Takes a data.frame as the starting point.

plotSurfaceComparison(
  object,
  of_sample = "",
  variables,
  method_gs = "mean",
  normalize = TRUE,
  smooth = FALSE,
  smooth_span = 0.02,
  pt_size = 2,
  pt_alpha = 1,
  pt_clrsp = "inferno",
  display_image = FALSE,
  assign = FALSE,
  assign_name,
  verbose = TRUE,
  ...
)

plotSurfaceComparison2(
  data,
  variables = "all",
  pt_size = 2,
  pt_alpha = 0.9,
  pt_clrsp = "inferno",
  image = NULL,
  verbose = TRUE,
  ...
)

Arguments

variables
Character vector. The variables of interest:

Gene sets: Must be in getGeneSets()

Genes: Must be in getGenes()

Features: Must be numeric ones of getFeatureNames()

method_gs
Character value. The method according to which gene sets will be handled specified as a character of length one. This can be either 'mean' or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().

smooth
Logical. If set to TRUE values will be smoothed according to the smoooth_-parameters.

smooth_span
The amount of smoothing specified as a single numeric value.

pt_size
The size of the points specified as a single numeric value.

pt_alpha
The transparency of the points specified as single numeric value.

pt_clrsp
The color spectrum to be used if the specified variable displayed by color is continuous. Run all_colorspectra() to see valid input..

display_image
Logical. If set to TRUE the histology image of the specified sample is displayed underneath the plot.

assign
Logical. If set to TRUE a named list will be assigned to the global environment. This list contains data and information to rebuild or additionally customize the output plot of this function.

assign_name
The name the assigned list is supposed to have specified as a single character value.

verbose
Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

...
Additional arguments given to ggplot2::facet_wrap().

data
A data.frame containing the numeric variables of interest.

image
An image of class Image to be displayed in the background. Easily accessible via SPATA::image().