Expression dynamic in heatmap

Displays variable-expression values along a trajectory direction with a smoothed heatmap (from left to right).

plotTrajectoryHeatmap(
  object,
  trajectory_name,
  of_sample = "",
  variables,
  method_gs = "mean",
  arrange_rows = "none",
  show_rownames = FALSE,
  show_colnames = FALSE,
  split_columns = TRUE,
  smooth_span = 0.5,
  verbose = TRUE,
  hm_colors = viridis::inferno(150),
  ...
)

Arguments

object
A valid spata-object.

trajectory_name
The trajectory of interest specified as a single character value.

of_sample
The sample(s) of interest specified as a single character value or vector. If set to "" (the default) the first sample is chosen.

variables
The variables of interest specified as a character vector:

Gene sets: Must be in getGeneSets()

Genes: Must be in getGenes()

All elements of the specified character vector must either belong to gene sets or to genes.

arrange_rows
Alter the way the rows of the heatmap are displayed in order to highlight patterns. Currently either 'maxima' or 'minima'.

show_rownames
Logical. If set to TRUE the variable elements will be displayed at the rownames of the heatmap.

split_columns
Logial. If set to TRUE the heatmap is vertically splitted according to the trajectory parts.

smooth_span
The amount of smoothing specified as a single numeric value.

verbose
Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

hm_colors
A vector of colors to be used.

...
Additional parameters given to pheatmap::pheatmap()

Value

A heatmap of class 'pheatmap'.